requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
dnspython==2.6.1
frozenlist==1.4.1
idna==3.6
MouseInfo==0.1.3
multidict==6.0.5
pillow==10.2.0
py-cord==2.5.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pymongo==4.6.2
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pytweening==1.2.0
pytz==2024.1
requests==2.31.0
schedule==1.2.1
urllib3==2.2.1
yarl==1.9.4
